refactor: split up endpoints.py into smaller modules
